refactor(ColorPicker): change component structure
